+++ src/wallet/api/wallet2_api.h
@@ -604,7 +604,7 @@ struct Wallet
     * @brief listCurrentStakes - returns a list of the wallets locked stakes, provides both service node address and the staked amount
     * @return
     */
-    virtual std::vector<std::pair<std::string, uint32_t>>* listCurrentStakes() const = 0;
+    virtual std::vector<std::pair<std::string, uint64_t>>* listCurrentStakes() const = 0;
 
    /**
     * @brief watchOnly - checks if wallet is watch only
